+++ packages/eyes-sdk-core/lib/EyesCore.js
@@ -9,6 +9,19 @@ const TestFailedError = require('./errors/TestFailedError')
 const EyesUtils = require('./EyesUtils')
 const EyesBase = require('./EyesBase')
 const Logger = require('./logging/Logger')
+const NullCutProvider = require('./cropping/NullCutProvider')
+const EyesScreenshot = require('./capture/EyesScreenshotNew')
+const GeneralUtils = require('./utils/GeneralUtils')
+const SimplePropertyHandler = require('./handler/SimplePropertyHandler')
+const NullScaleProvider = require('./scaling/NullScaleProvider')
+const ScaleProviderIdentityFactory = require('./scaling/ScaleProviderIdentityFactory')
+const FixedScaleProviderFactory = require('./scaling/FixedScaleProviderFactory')
+const ContextBasedScaleProviderFactory = require('./scaling/ContextBasedScaleProviderFactory')
+const UserAgent = require('./useragent/UserAgent')
+const ImageProviderFactory = require('./capture/ImageProviderFactory')
+
+const UNKNOWN_DEVICE_PIXEL_RATIO = 0
+const DEFAULT_DEVICE_PIXEL_RATIO = 1
 
 /**
  * @template TDriver, TElement, TSelector
@@ -31,6 +44,45 @@ const Logger = require('./logging/Logger')
  * @template TSelector
  */
 class EyesCore extends EyesBase {
+  constructor(serverUrl, isDisabled) {
+    super(serverUrl, isDisabled)
+
+    /** @type {EyesWrappedDriver<TDriver, TElement, TSelector>} */
+    this._driver = undefined
+    /** @private @type {EyesJsExecutor<TDriver, TElement, TSelector>} */
+    this._executor = undefined
+    /** @private @type {EyesElementFinder<TDriver, TElement, TSelector>} */
+    this._finder = undefined
+    /** @private @type {EyesBrowsingContext<TDriver, TElement, TSelector>} */
+    this._context = undefined
+    /** @private @type {EyesDriverController<TDriver, TElement, TSelector>} */
+    this._controller = undefined
+    /** @private @type {boolean} */
+    this._dontGetTitle = false
+
+    /** @private @type {number} */
+    this._devicePixelRatio = UNKNOWN_DEVICE_PIXEL_RATIO
+    /** @private */
+    this._rotation = undefined
+  }
+
+  async _initCommon() {
+    this._devicePixelRatio = UNKNOWN_DEVICE_PIXEL_RATIO
+
+    const userAgentString = await this._controller.getUserAgent()
+    if (userAgentString) {
+      this._userAgent = UserAgent.parseUserAgentString(userAgentString, true)
+    }
+
+    this._imageProvider = ImageProviderFactory.getImageProvider(
+      this._logger,
+      this._driver,
+      this._rotation,
+      this,
+      this._userAgent,
+    )
+  }
+
   /* ------------ Classic API ------------ */
   /**
    * Takes a snapshot of the application under test and matches it with the expected output.
@@ -338,6 +390,134 @@ class EyesCore extends EyesBase {
 
     this._viewportSizeHandler.set(new RectangleSize(viewportSize))
   }
+
+  /**
+   * Run visual locators
+   * @param visualLocatorSettings
+   * @param {string[]} visualLocatorSettings.locatorNames
+   * @param {boolean} visualLocatorSettings.firstOnly
+   * @return {Promise<{[key: string]: import('../typings/lib/geometry/Region').RegionObject[]}>}
+   */
+  async locate(visualLocatorSettings) {
+    ArgumentGuard.notNull(visualLocatorSettings, 'visualLocatorSettings')
+    this._logger.verbose('Get locators with given names: ', visualLocatorSettings.locatorNames)
+    const screenshot = await this._getViewportScreenshot()
+    const screenshotBuffer = await screenshot.getImage().getImageBuffer()
+    const id = GeneralUtils.guid()
+    await this.getAndSaveRenderingInfo()
+    const imageUrl = await this._serverConnector.uploadScreenshot(id, screenshotBuffer)
+    const appName = this._configuration.getAppName()
+    return this._serverConnector.postLocators({
+      appName,
+      imageUrl,
+      locatorNames: visualLocatorSettings.locatorNames,
+      firstOnly: visualLocatorSettings.firstOnly,
+    })
+  }
+
+  /**
+   * Create a viewport page screenshot
+   * @return {Promise<EyesScreenshot>}
+   */
+  async _getViewportScreenshot() {
+    this._logger.verbose('Screenshot requested...')
+    const scaleProviderFactory = await this._updateScalingParams()
+
+    let screenshotImage = await this._imageProvider.getImage()
+    await this._debugScreenshotsProvider.save(screenshotImage, 'original')
+
+    const scaleProvider = scaleProviderFactory.getScaleProvider(screenshotImage.getWidth())
+    if (scaleProvider.getScaleRatio() !== 1) {
+      this._logger.verbose('scaling...')
+      screenshotImage = await screenshotImage.scale(scaleProvider.getScaleRatio())
+      await this._debugScreenshotsProvider.save(screenshotImage, 'scaled')
+    }
+
+    const cutProvider = this._cutProviderHandler.get()
+    if (!(cutProvider instanceof NullCutProvider)) {
+      this._logger.verbose('cutting...')
+      screenshotImage = await cutProvider.cut(screenshotImage)
+      await this._debugScreenshotsProvider.save(screenshotImage, 'cut')
+    }
+
+    this._logger.verbose('Building screenshot object...')
+    return EyesScreenshot.fromScreenshotType(this._logger, this, screenshotImage)
+  }
+
+  /**
+   * @private
+   * @return {Promise<ScaleProviderFactory>}
+   */
+  async _updateScalingParams() {
+    // Update the scaling params only if we haven't done so yet, and the user hasn't set anything else manually.
+    if (
+      this._devicePixelRatio !== UNKNOWN_DEVICE_PIXEL_RATIO &&
+      !(this._scaleProviderHandler.get() instanceof NullScaleProvider)
+    ) {
+      // If we already have a scale provider set, we'll just use it, and pass a mock as provider handler.
+      const nullProvider = new SimplePropertyHandler()
+      return new ScaleProviderIdentityFactory(this._scaleProviderHandler.get(), nullProvider)
+    }
+
+    this._logger.verbose('Trying to extract device pixel ratio...')
+    this._devicePixelRatio = await EyesUtils.getDevicePixelRatio(this._logger, this._driver)
+      .catch(async err => {
+        const isNative = await this._controller.isNative()
+        if (!isNative) throw err
+        const viewportSize = await this.getViewportSize()
+        return EyesUtils.getMobilePixelRatio(this._logger, this._driver, viewportSize)
+      })
+      .catch(err => {
+        this._logger.verbose('Failed to extract device pixel ratio! Using default.', err)
+        return DEFAULT_DEVICE_PIXEL_RATIO
+      })
+
+    this._logger.verbose(`Device pixel ratio: ${this._devicePixelRatio}`)
+    this._logger.verbose('Setting scale provider...')
+    const factory = await this._getScaleProviderFactory().catch(err => {
+      this._logger.verbose('Failed to set ContextBasedScaleProvider.', err)
+      this._logger.verbose('Using FixedScaleProvider instead...')
+      return new FixedScaleProviderFactory(1 / this._devicePixelRatio, this._scaleProviderHandler)
+    })
+    this._logger.verbose('Done!')
+    return factory
+  }
+
+  /**
+   * @private
+   * @return {Promise<ScaleProviderFactory>}
+   */
+  async _getScaleProviderFactory() {
+    const entireSize = await EyesUtils.getCurrentFrameContentEntireSize(
+      this._logger,
+      this._executor,
+    )
+    return new ContextBasedScaleProviderFactory(
+      this._logger,
+      entireSize,
+      this._viewportSizeHandler.get(),
+      this._devicePixelRatio,
+      false,
+      this._scaleProviderHandler,
+    )
+  }
+
+  /**
+   * @private
+   */
+  async _getAndSaveBatchInfoFromServer(batchId) {
+    ArgumentGuard.notNullOrEmpty(batchId, 'batchId')
+    return this._runner.getBatchInfoWithCache(batchId)
+  }
+
+  /**
+   * @private
+   */
+  async getAndSaveRenderingInfo() {
+    const renderingInfo = await this._runner.getRenderingInfoWithCache()
+    this._serverConnector.setRenderingInfo(renderingInfo)
+  }
+
   async getAUTSessionId() {
     if (!this._driver) {
       return undefined
