+++ packages/sdk-coverage-tests/src/supported-tests.js
@@ -311,4 +311,6 @@ module.exports = [
   {name: 'TestWindowWithModal_Fully', executionMode: {isScrollStitching: true}},
   {name: 'TestTooBigViewportSize', executionMode: {isCssStitching: true}},
   {name: 'TestSetViewportSize', executionMode: {isCssStitching: true}},
+  {name: 'TestVisualLocators', executionMode: {isCssStitching: true}},
+  {name: 'TestVisualLocators', executionMode: {isVisualGrid: true}},
 ]
